+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## [vnext]
 ### Fixed
 - Target.region() outside of viewport on MobileSafari with iOS Simulator doesn't work correctly [Trello 1708](https://trello.com/c/wrMoRDPY)
+- Parsing CSS has failed if `@charset` was present [Trello 2227](https://trello.com/c/2AjLHPFw)
 ### Updated
+- Abort UFG test if dom-snapshot script failed [Trello 2227](https://trello.com/c/2AjLHPFw)
 - Lifted pillow restriction to support python 3.9 [Trello 2303](https://trello.com/c/EcpNN2tY)
 
 ## [4.11.3] - 2020-11-03